requirements.txt
langchain
langchain-community
langchainhub
langchain-openai
langchain-chroma
bs4
python-dotenv
pypdf2
python-multipart